Consolidate CurationConcernController and WorksControllerBehavior
This was a remaining step in merging Sufia and CurationConcerns
