+++ app/controllers/concerns/hyrax/works_controller_behavior.rb
@@ -1,11 +1,42 @@
 module Hyrax
   module WorksControllerBehavior
     extend ActiveSupport::Concern
-    # TODO: merge with CurationConcernController
-    include Hyrax::CurationConcernController
+    include Blacklight::Base
+    include Blacklight::AccessControls::Catalog
 
     included do
       layout :decide_layout
+
+      copy_blacklight_config_from(::CatalogController)
+
+      class_attribute :_curation_concern_type, :show_presenter, :work_form_service, :search_builder_class
+      self.show_presenter = Hyrax::WorkShowPresenter
+      self.work_form_service = Hyrax::WorkFormService
+      self.search_builder_class = WorkSearchBuilder
+      attr_accessor :curation_concern
+      helper_method :curation_concern, :contextual_path
+
+      rescue_from WorkflowAuthorizationException, with: :render_unavailable
+    end
+
+    class_methods do
+      def curation_concern_type=(curation_concern_type)
+        load_and_authorize_resource class: curation_concern_type, instance_name: :curation_concern, except: [:show, :file_manager, :inspect_work]
+
+        # Load the fedora resource to get the etag.
+        # No need to authorize for the file manager, because it does authorization via the presenter.
+        load_resource class: curation_concern_type, instance_name: :curation_concern, only: :file_manager
+
+        self._curation_concern_type = curation_concern_type
+      end
+
+      def curation_concern_type
+        _curation_concern_type
+      end
+
+      def cancan_resource_class
+        Hyrax::ControllerResource
+      end
     end
 
     def new
@@ -17,10 +48,164 @@ module Hyrax
       # Set the first admin_set they have access to.
       admin_set = Hyrax::AdminSetService.new(self).search_results(:deposit).first
       curation_concern.admin_set_id = admin_set && admin_set.id
-      super
+      build_form
     end
 
-    protected
+    def create
+      if actor.create(attributes_for_actor)
+        after_create_response
+      else
+        respond_to do |wants|
+          wants.html do
+            build_form
+            render 'new', status: :unprocessable_entity
+          end
+          wants.json { render_json_response(response_type: :unprocessable_entity, options: { errors: curation_concern.errors }) }
+        end
+      end
+    end
+
+    # Finds a solr document matching the id and sets @presenter
+    # @raise CanCan::AccessDenied if the document is not found or the user doesn't have access to it.
+    def show
+      respond_to do |wants|
+        wants.html { presenter && parent_presenter }
+        wants.json do
+          # load and authorize @curation_concern manually because it's skipped for html
+          @curation_concern = _curation_concern_type.find(params[:id]) unless curation_concern
+          authorize! :show, @curation_concern
+          render :show, status: :ok
+        end
+        additional_response_formats(wants)
+        wants.ttl do
+          render body: presenter.export_as_ttl, content_type: 'text/turtle'
+        end
+        wants.jsonld do
+          render body: presenter.export_as_jsonld, content_type: 'application/ld+json'
+        end
+        wants.nt do
+          render body: presenter.export_as_nt, content_type: 'application/n-triples'
+        end
+      end
+    end
+
+    def edit
+      build_form
+    end
+
+    def update
+      if actor.update(attributes_for_actor)
+        after_update_response
+      else
+        respond_to do |wants|
+          wants.html do
+            build_form
+            render 'edit', status: :unprocessable_entity
+          end
+          wants.json { render_json_response(response_type: :unprocessable_entity, options: { errors: curation_concern.errors }) }
+        end
+      end
+    end
+
+    def destroy
+      title = curation_concern.to_s
+      return unless actor.destroy
+      Hyrax.config.callback.run(:after_destroy, curation_concern.id, current_user)
+      after_destroy_response(title)
+    end
+
+    def file_manager
+      @form = Forms::FileManagerForm.new(curation_concern, current_ability)
+    end
+
+    def inspect_work
+      raise Hydra::AccessDenied unless current_ability.admin?
+      presenter
+    end
+
+    #    attr_writer :actor
+
+    private
+
+      def build_form
+        @form = work_form_service.build(curation_concern, current_ability, self)
+      end
+
+      def actor
+        @actor ||= Hyrax::CurationConcern.actor(curation_concern, current_ability)
+      end
+
+      def presenter
+        @presenter ||= show_presenter.new(curation_concern_from_search_results, current_ability, request)
+      end
+
+      def parent_presenter
+        @parent_presenter ||=
+          begin
+            if params[:parent_id]
+              @parent_presenter ||= show_presenter.new(search_result_document(id: params[:parent_id]), current_ability, request)
+            end
+          end
+      end
+
+      # Include 'hyrax/base' in the search path for views, while prefering
+      # our local paths. Thus we are unable to just override `self.local_prefixes`
+      def _prefixes
+        @_prefixes ||= super + ['hyrax/base']
+      end
+
+      def hash_key_for_curation_concern
+        _curation_concern_type.model_name.param_key
+      end
+
+      def contextual_path(presenter, parent_presenter)
+        ::Hyrax::ContextualPath.new(presenter, parent_presenter).show
+      end
+
+      def curation_concern_from_search_results
+        search_result_document(params)
+      end
+
+      # Only returns unsuppressed documents the user has read access to
+      def search_result_document(search_params)
+        _, document_list = search_results(search_params)
+        return document_list.first unless document_list.empty?
+        document_not_found!
+      end
+
+      def document_not_found!
+        doc = ::SolrDocument.find(params[:id])
+        raise WorkflowAuthorizationException if doc.suppressed?
+        raise CanCan::AccessDenied.new(nil, :show)
+      end
+
+      def render_unavailable
+        message = I18n.t("hyrax.workflow.unauthorized")
+        respond_to do |wants|
+          wants.html do
+            unavailable_presenter
+            flash[:notice] = message
+            render 'unavailable', status: :unauthorized
+          end
+          wants.json do
+            render plain: message, status: :unauthorized
+          end
+          additional_response_formats(wants)
+          wants.ttl do
+            render plain: message, status: :unauthorized
+          end
+          wants.jsonld do
+            render plain: message, status: :unauthorized
+          end
+          wants.nt do
+            render plain: message, status: :unauthorized
+          end
+        end
+      end
+
+      def unavailable_presenter
+        @presenter ||= show_presenter.new(::SolrDocument.find(params[:id]), current_ability, request)
+      end
 
       def decide_layout
         case action_name
@@ -33,7 +218,13 @@ module Hyrax
 
       # Add uploaded_files to the parameters received by the actor.
       def attributes_for_actor
-        attributes = super
+        raw_params = params[hash_key_for_curation_concern]
+        attributes = if raw_params
+                       work_form_service.form_class(curation_concern).model_attributes(raw_params)
+                     else
+                       {}
+                     end
+
         # If they selected a BrowseEverything file, but then clicked the
         # remove button, it will still show up in `selected_files`, but
         # it will no longer be in uploaded_files. By checking the
@@ -66,10 +257,17 @@ module Hyrax
       end
 
       def after_update_response
-        if permissions_changed? && curation_concern.file_sets.present?
-          redirect_to hyrax.confirm_access_permission_path(curation_concern)
+        if curation_concern.file_sets.present?
+          if permissions_changed?
+            redirect_to hyrax.confirm_access_permission_path(curation_concern)
+          elsif curation_concern.visibility_changed?
+            redirect_to main_app.confirm_hyrax_permission_path(curation_concern)
+          end
         else
-          super
+          respond_to do |wants|
+            wants.html { redirect_to [main_app, curation_concern] }
+            wants.json { render :show, status: :ok, location: polymorphic_path([main_app, curation_concern]) }
+          end
         end
       end
 
@@ -80,7 +278,6 @@ module Hyrax
         end
       end
 
-      # Called by Hyrax::CurationConcernController#show
       def additional_response_formats(format)
         format.endnote do
           send_data(presenter.solr_document.export_as_endnote,
